+++ examples/json_parser_auto/src/json_grammar_trait.rs
@@ -10,12 +10,11 @@
 #![allow(clippy::upper_case_acronyms)]
 
 use parol_runtime::derive_builder::Builder;
-use parol_runtime::lexer::Token;
 use parol_runtime::log::trace;
 #[allow(unused_imports)]
 use parol_runtime::parol_macros::{pop_and_reverse_item, pop_item};
 use parol_runtime::parser::{ParseTreeType, UserActionsTrait};
-use parol_runtime::{ParserError, Result};
+use parol_runtime::{ParserError, Result, Token};
 
 /// Semantic actions trait generated for the user grammar
 /// All functions have default implementations.
@@ -835,4 +834,5 @@ impl<'t> UserActionsTrait<'t> for JsonGrammarAuto<'t, '_> {
             .into()),
         }
     }
+    fn on_comment_parsed(&mut self, _token: Token<'t>) {}
 }