+++ crates/parol_runtime/CHANGELOG.md
@@ -16,6 +16,13 @@ Be aware that this project is still v0.y.z which means that anything can change 
 We defined for this project that while being on major version zero we mark incompatible changes with
 new minor version numbers. Please note that this is no version handling covered by `Semver`.
 
+## 0.17.0 - Not released yet
+
+* New support for handling of user defined comments (`%line_comment`, `%block_comment`)
+  * The new method `UserActionsTrait::on_comment_parsed` is called in order of appearance each time
+  before the parser consumes a normal token from token stream.
+  * It is still a bit under investigation where best call this method during parse process.
+
 ## 0.16.0 - 2023-04-02
 
 * More efficient implementation of lookahead DFA